Now replies to Tweets with the name of the song.

+++ naturedroid.py
@@ -4,15 +4,19 @@ import os
 from moviepy.editor import *
 from twython import Twython
 
-APP_KEY = ''
-APP_SECRET = ''
-OAUTH_TOKEN = ''
-OAUTH_SECRET = ''
+#APP_KEY = ''
+#APP_SECRET = ''
+#OAUTH_TOKEN = ''
+#OAUTH_SECRET = ''
+#ACCOUNT_ID = ''
 
 twitter = Twython(APP_KEY, APP_SECRET, OAUTH_TOKEN, OAUTH_SECRET)
 
 # Length of desired clip (in seconds)
 CLIP_LENGTH = 30
+
+# Name of song for the current clip
+song_name = ""
 
 # Randomly selects a video and grabs a sublcip from it
 def get_video(list_of_docs):
@@ -29,9 +33,11 @@ def get_video(list_of_docs):
 # Randomly selects a song and grabs a snippet from it
 def get_song(list_of_songs):
     global CLIP_LENGTH
+    global song_name
     song_num = random.randint(0, len(list_of_songs) - 1)
     song = AudioFileClip("songs/" + list_of_songs[song_num])
     print("Chose song: " + list_of_songs[song_num])
+    song_name = list_of_songs[song_num].replace(".mp3", "")
     song_length = math.floor(song.duration)
     start_time = random.randint(0, song_length - CLIP_LENGTH)
     mp3 = song.subclip(start_time, start_time + CLIP_LENGTH)
@@ -44,12 +50,16 @@ def create_clip(list_of_songs, list_of_docs):
     clip = get_video(list_of_docs)
     clip = clip.set_audio(mp3)
     return clip
-
+    
 # Main function
 def main():
+    global song_name
     print("Starting...")
     song_list = os.listdir("songs/")
     doc_list = os.listdir("docs/")
+    emoji_file = open("emoji_list.txt", "r")
+    emojis = emoji_file.read().split(",")
+    emoji_file.close()
     print("Generating clip...")
     clip = create_clip(song_list, doc_list)
     print("Writing video...")
@@ -61,7 +71,14 @@ def main():
     response = twitter.upload_video(media=video, media_type='video/mp4')
     twitter.update_status(status='', media_ids=[response['media_id']])
     print("Done! Video has been tweeted.")
-    os.remove("clip.mp4")
+    new_tweet = twitter.get_user_timeline(user_id = ACCOUNT_ID, count_id = 1)[0]
+    screen_name = new_tweet['user']['screen_name']
+    tweet_id = new_tweet['id']
+    random_emoji = emojis[random.randint(0, len(emojis) - 1)].encode('utf-8').decode('unicode-escape')
+    twitter.update_status(status = "@" + screen_name + " " + random_emoji + " song: " + song_name,
+                          in_reply_to_status_id = tweet_id)
+    if (os.path.exists("clip.mp4")):
+        os.remove("clip.mp4")
 
 # Starts the main function
 main()
